fix: widen base64url capacity math to prevent int overflow (R-145)
ss_base64url_encode computed `((input_byte_count * 4) + 2) / 3 + 1` in signed
int before the capacity check, so a large input_byte_count (> ~536M) overflowed
into a small/negative required_capacity that slipped past
`output_buffer_capacity < required_capacity` and overflowed the output buffer. It
now computes required_capacity in long long and compares
(long long)output_buffer_capacity, so the math can't overflow and a genuinely
too-large input fails the capacity check cleanly (SS_BCRYPT_ERR_CONFIG).

The bcrypt runtime rebuilds and taskforge-web's auth (bcrypt hash + base64url
session token) round-trips (all 7 apps pass). Triggering needs a >0.5GB input,
so a source-guard test keeps the widened math in place. pytest 1104 green.

+++ semanticscript/runtime/native_bcrypt/sem_bcrypt_runtime.c
@@ -233,8 +233,13 @@ int ss_base64url_encode(
     if (input_buffer == NULL || output_buffer == NULL || input_byte_count < 0) {
         return SS_BCRYPT_ERR_CONFIG;
     }
-    int required_capacity = ((input_byte_count * 4) + 2) / 3 + 1; /* +1 for NUL */
-    if (output_buffer_capacity < required_capacity) {
+    /* R-145: compute the required capacity in a wide type so `input_byte_count *
+     * 4` cannot overflow signed int — an overflow would produce a small/negative
+     * capacity that slips past the bounds check below and overflows the output
+     * buffer. A genuinely too-large input then fails the capacity check cleanly. */
+    long long required_capacity =
+        (((long long)input_byte_count * 4) + 2) / 3 + 1; /* +1 for NUL */
+    if ((long long)output_buffer_capacity < required_capacity) {
         return SS_BCRYPT_ERR_CONFIG;
     }
